fix: move the ClientSplashPresenter to the right module
