+++ shared/presentation/src/commonMain/kotlin/network/bisq/mobile/presentation/di/PresentationModule.kt
@@ -61,8 +61,10 @@ import network.bisq.mobile.presentation.ui.uicases.settings.ResourcesPresenter
 import network.bisq.mobile.presentation.ui.uicases.settings.SettingsPresenter
 import network.bisq.mobile.presentation.ui.uicases.settings.SupportPresenter
 import network.bisq.mobile.presentation.ui.uicases.settings.UserProfilePresenter
+import network.bisq.mobile.presentation.ui.uicases.startup.ClientSplashPresenter
 import network.bisq.mobile.presentation.ui.uicases.startup.CreateProfilePresenter
 import network.bisq.mobile.presentation.ui.uicases.startup.IAgreementPresenter
+import network.bisq.mobile.presentation.ui.uicases.startup.SplashPresenter
 import network.bisq.mobile.presentation.ui.uicases.startup.TrustedNodeSetupPresenter
 import network.bisq.mobile.presentation.ui.uicases.startup.UserAgreementPresenter
 import network.bisq.mobile.presentation.ui.uicases.take_offer.TakeOfferAmountPresenter
@@ -96,6 +98,17 @@ val presentationModule = module {
     } bind AppPresenter::class
 
     single<BannerPresenter> { BannerPresenter(get(), get()) }
+
+    single<SplashPresenter> {
+        ClientSplashPresenter(
+            get(),
+            get(),
+            get(),
+            get(),
+            get(),
+            get(),
+        )
+    }
 
     single<TopBarPresenter> { TopBarPresenter(get(), get(), get(), get()) } bind ITopBarPresenter::class
